1.支持AND,OR 多条件叠加查询 2.支持CURD

+++ src/main/java/org/beetl/sql/core/query/interfacer/QueryOtherI.java
@@ -38,6 +38,6 @@ public interface QueryOtherI<T extends QueryOtherI>{
      * @param startRow
      * @return
      */
-    T limit(Long startRow,Long rowCount);
+    T limit(Integer startRow,Integer rowCount);
 
 }
